access-checkout-react-native-sdk/android/src/main/java/com/worldpay/access/checkout/reactnative/AccessCheckoutReactNativePackage.kt
package com.worldpay.access.checkout.reactnative

import com.facebook.react.ReactPackage
import com.facebook.react.bridge.NativeModule
import com.facebook.react.bridge.ReactApplicationContext
import com.facebook.react.uimanager.ViewManager

/**
 * This class is responsible for registering native modules with React Native. (called upon initialisation)
 */
class AccessCheckoutReactNativePackage : ReactPackage {

    override fun createNativeModules(reactContext: ReactApplicationContext): List<NativeModule> {
        return listOf(AccessCheckoutReactNativeModule(reactContext))
    }

    override fun createViewManagers(reactContext: ReactApplicationContext): List<ViewManager<*, *>> {
        return emptyList()
    }
}
